Added more operations

+++ src/main/java/com/bernardomg/association/member/repository/MemberRepository.java
@@ -28,6 +28,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bernardomg.association.member.model.PersistentMember;
 
-public interface MemberRepository extends JpaRepository<PersistentMember, Integer> {
+public interface MemberRepository extends JpaRepository<PersistentMember, Long> {
 
 }
